+++ pkg/bpmn_engine/actiivty_test.go
@@ -1,0 +1,5 @@
+package bpmn_engine
+
+//func Test_Activity_interfaces_implemented(t *testing.T) {
+//	var _ Activity = &activity{}
+//}