+++ pkg/bpmn_engine/engine_jobs_test.go
@@ -6,8 +6,6 @@ import (
 	"github.com/corbym/gocrest/has"
 	"github.com/corbym/gocrest/is"
 	"github.com/corbym/gocrest/then"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/activity"
-	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/process_instance"
 )
 
 const (
@@ -31,6 +29,10 @@ func jobCompleteHandler(job ActivatedJob) {
 	job.Complete()
 }
 
+func Test_job_implements_Activity(t *testing.T) {
+	var _ Activity = &job{}
+}
+
 func Test_a_job_can_fail_and_keeps_the_instance_in_active_state(t *testing.T) {
 	// setup
 	bpmnEngine := New()
@@ -39,7 +41,7 @@ func Test_a_job_can_fail_and_keeps_the_instance_in_active_state(t *testing.T) {
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.State, is.EqualTo(Active))
 }
 
 // Test_simple_count_loop requires correct Task-Output-Mapping in the BPMN file
@@ -54,7 +56,7 @@ func Test_simple_count_loop(t *testing.T) {
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, vars)
 
 	then.AssertThat(t, instance.GetVariable(varCounter), is.EqualTo(4))
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 }
 
 func Test_simple_count_loop_with_message(t *testing.T) {
@@ -85,12 +87,12 @@ func Test_simple_count_loop_with_message(t *testing.T) {
 
 	then.AssertThat(t, instance.GetVariable(varHasReachedMaxAttempts), is.True())
 	then.AssertThat(t, instance.GetVariable(varEngineValidationAttempts), is.EqualTo(2))
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 
 	// internal State expected
 	then.AssertThat(t, bpmnEngine.GetMessageSubscriptions(), has.Length(2))
-	then.AssertThat(t, bpmnEngine.GetMessageSubscriptions()[0].State, is.EqualTo(activity.Completed))
-	then.AssertThat(t, bpmnEngine.GetMessageSubscriptions()[1].State, is.EqualTo(activity.Completed))
+	then.AssertThat(t, bpmnEngine.GetMessageSubscriptions()[0].State, is.EqualTo(Completed))
+	then.AssertThat(t, bpmnEngine.GetMessageSubscriptions()[1].State, is.EqualTo(Completed))
 }
 
 func Test_activated_job_data(t *testing.T) {
@@ -108,7 +110,7 @@ func Test_activated_job_data(t *testing.T) {
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
 
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.State, is.EqualTo(Active))
 }
 
 func Test_task_InputOutput_mapping_happy_path(t *testing.T) {
@@ -118,8 +120,8 @@ func Test_task_InputOutput_mapping_happy_path(t *testing.T) {
 
 	// give
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/service-task-input-output.bpmn")
-	bpmnEngine.NewTaskHandler().Id("service-task-1").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("user-task-2").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("service-task-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("user-task-2").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -127,7 +129,7 @@ func Test_task_InputOutput_mapping_happy_path(t *testing.T) {
 
 	// then
 	for _, job := range bpmnEngine.jobs {
-		then.AssertThat(t, job.State, is.EqualTo(activity.Completed))
+		then.AssertThat(t, job.JobState, is.EqualTo(Completed))
 	}
 	then.AssertThat(t, cp.CallPath, is.EqualTo("service-task-1,user-task-2"))
 	// id from input should not exist in instance scope
@@ -149,7 +151,7 @@ func Test_instance_fails_on_Invalid_Input_mapping(t *testing.T) {
 
 	// give
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/service-task-invalid-input.bpmn")
-	bpmnEngine.NewTaskHandler().Id("invalid-input").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("invalid-input").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -158,8 +160,8 @@ func Test_instance_fails_on_Invalid_Input_mapping(t *testing.T) {
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo(""))
 	then.AssertThat(t, pi.GetVariable("id"), is.Nil())
-	then.AssertThat(t, bpmnEngine.jobs[0].State, is.EqualTo(activity.Failed))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.FAILED))
+	then.AssertThat(t, bpmnEngine.jobs[0].JobState, is.EqualTo(Failed))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Failed))
 }
 
 func Test_job_fails_on_Invalid_Output_mapping(t *testing.T) {
@@ -169,7 +171,7 @@ func Test_job_fails_on_Invalid_Output_mapping(t *testing.T) {
 
 	// give
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/service-task-invalid-output.bpmn")
-	bpmnEngine.NewTaskHandler().Id("invalid-output").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("invalid-output").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -178,8 +180,8 @@ func Test_job_fails_on_Invalid_Output_mapping(t *testing.T) {
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("invalid-output"))
 	then.AssertThat(t, pi.GetVariable("order"), is.Nil())
-	then.AssertThat(t, bpmnEngine.jobs[0].State, is.EqualTo(activity.Failed))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.FAILED))
+	then.AssertThat(t, bpmnEngine.jobs[0].JobState, is.EqualTo(Failed))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Failed))
 }
 
 func Test_task_type_handler(t *testing.T) {
@@ -189,7 +191,7 @@ func Test_task_type_handler(t *testing.T) {
 
 	// give
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/simple-task-with-type.bpmn")
-	bpmnEngine.NewTaskHandler().Type("foobar").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Type("foobar").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -197,7 +199,7 @@ func Test_task_type_handler(t *testing.T) {
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("id"))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_task_type_handler_ID_handler_has_precedence(t *testing.T) {
@@ -224,7 +226,7 @@ func Test_task_type_handler_ID_handler_has_precedence(t *testing.T) {
 
 	// then
 	then.AssertThat(t, calledHandler, is.EqualTo("ID"))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_just_one_handler_called(t *testing.T) {
@@ -234,9 +236,9 @@ func Test_just_one_handler_called(t *testing.T) {
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/simple-task-with-type.bpmn")
 
 	// given multiple matching handlers executed
-	bpmnEngine.NewTaskHandler().Id("id").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Id("id").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().Type("foobar").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Id("id").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Type("foobar").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -244,7 +246,7 @@ func Test_just_one_handler_called(t *testing.T) {
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("id").Reason("just one execution"))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_assignee_and_candidate_groups_are_assigned_to_handler(t *testing.T) {
@@ -254,8 +256,8 @@ func Test_assignee_and_candidate_groups_are_assigned_to_handler(t *testing.T) {
 	process, _ := bpmnEngine.LoadFromFile("../../test-cases/user-tasks-with-assignments.bpmn")
 
 	// given multiple matching handlers executed
-	bpmnEngine.NewTaskHandler().Assignee("john.doe").Handler(cp.CallPathHandler)
-	bpmnEngine.NewTaskHandler().CandidateGroups("marketing", "support").Handler(cp.CallPathHandler)
+	bpmnEngine.NewTaskHandler().Assignee("john.doe").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().CandidateGroups("marketing", "support").Handler(cp.TaskHandler)
 
 	// when
 	pi, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
@@ -263,7 +265,7 @@ func Test_assignee_and_candidate_groups_are_assigned_to_handler(t *testing.T) {
 
 	// then
 	then.AssertThat(t, cp.CallPath, is.EqualTo("assignee-task,group-task"))
-	then.AssertThat(t, pi.GetState(), is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, pi.GetState(), is.EqualTo(Completed))
 }
 
 func Test_task_default_all_output_variables_map_to_process_instance(t *testing.T) {
@@ -276,7 +278,7 @@ func Test_task_default_all_output_variables_map_to_process_instance(t *testing.T
 	})
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 
 	then.AssertThat(t, instance.GetVariable("aVariable"), is.True())
 }
@@ -291,7 +293,7 @@ func Test_task_no_output_variables_mapping_on_failure(t *testing.T) {
 	})
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.ACTIVE))
+	then.AssertThat(t, instance.State, is.EqualTo(Active))
 
 	then.AssertThat(t, instance.GetVariable("aVariable"), is.Nil())
 }
@@ -307,8 +309,33 @@ func Test_task_just_declared_output_variables_map_to_process_instance(t *testing
 	})
 
 	instance, _ := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
-	then.AssertThat(t, instance.State, is.EqualTo(process_instance.COMPLETED))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
 
 	then.AssertThat(t, instance.GetVariable("valueFromHandler"), is.True())
 	then.AssertThat(t, instance.GetVariable("otherVariable"), is.Nil())
+}
+
+func Test_missing_task_handlers_break_execution_and_can_be_continued_later(t *testing.T) {
+	cp := CallPath{}
+	// setup
+	bpmnEngine := New()
+	process, _ := bpmnEngine.LoadFromFile("../../test-cases/parallel-gateway-flow.bpmn")
+
+	// given
+	bpmnEngine.NewTaskHandler().Id("id-a-1").Handler(cp.TaskHandler)
+	instance, err := bpmnEngine.CreateAndRunInstance(process.ProcessKey, nil)
+	then.AssertThat(t, err, is.Nil())
+	then.AssertThat(t, instance.State, is.EqualTo(Active))
+	then.AssertThat(t, cp.CallPath, is.EqualTo("id-a-1"))
+
+	// when
+	bpmnEngine.NewTaskHandler().Id("id-b-1").Handler(cp.TaskHandler)
+	bpmnEngine.NewTaskHandler().Id("id-b-2").Handler(cp.TaskHandler)
+	instance, err = bpmnEngine.RunOrContinueInstance(instance.GetInstanceKey())
+	then.AssertThat(t, instance, is.Not(is.Nil()))
+	then.AssertThat(t, instance.State, is.EqualTo(Completed))
+
+	// then
+	then.AssertThat(t, err, is.Nil())
+	then.AssertThat(t, cp.CallPath, is.EqualTo("id-a-1,id-b-1,id-b-2"))
 }