+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * add feature to export and import BPMN state, incl. resume capability (#12)
 * make explicit engine name optional (#73 BREAKING CHANGE)
 * use global ID generator internally, to avoid ID collisions between multiple engine instances 
+* refactor `activity.LifecylceState` (BREAKING CHANGE)
+* refactor `process_instance.State` (BREAKING CHANGE)
 
 ### Migration notes for breaking changes